requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cssselect==1.3.0
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.124.4
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.55.0
google-search-results==2.4.2
googleapis-common-protos==1.72.0
groq==0.37.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.50.0
tavily-python==0.7.14
tenacity==9.1.2
tiktoken==0.12.0
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.38.0
websockets==15.0.1
Werkzeug==3.1.4
